+++ Changelog.md
@@ -13,6 +13,7 @@
 - Added shipment event classes (created, delivered, ready, lost, etc)
 - Added support to media settings `background` and `format` (passing "original" calls `keepOriginalFormat()`)
 - Added the `getDrivers()`, `choices()` and `ids()` static methods to the `TaxEngineManager` class
+- Added the `TaxEngineManager::dropResolvedInstances()` method
 - Added `Registerable` to all tax engine drivers (using the [konekt/xtend](https://github.com/artkonekt/xtend) interface)
 - Added the passing of `vanilo.taxes.engine.use_shipping_address` config flag to the Simple Tax Engine driver's constructor
 - Added the registration of `promotion` entry to the Eloquent morph map