+++ src/Taxes/Changelog.md
@@ -6,6 +6,7 @@
 ##### 2025-XX-YY
 
 - Added the `getDrivers()`, `choices()` and `ids()` static methods to the `TaxEngineManager` class
+- Added the `TaxEngineManager::dropResolvedInstances()` method
 - Added `Registerable` to all tax engine drivers (using the [konekt/xtend](https://github.com/artkonekt/xtend) interface)
 - Added the passing of `vanilo.taxes.engine.use_shipping_address` config flag to the Simple Tax Engine driver's constructor